// backend/template/src/app/services/models/meeting.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { User } from '../models/user';
export interface Meeting {
  admin?: User;
  beginDate?: string;
  createdAt?: string;
  description?: string;
  endDate?: string;
  formattedDate?: string;
  formattedTime?: string;
  meetingId?: number;
  members?: Array<User>;
  title?: string;
  updatedAt?: string;
}